Small coverage validation refactor
Related #3272

commit-id:075b62ba
